requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.2.0
cached-property==1.5.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.1
Flask==2.2.5
Flask-Cors==4.0.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.7
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
openai==0.28.0
pydantic==2.5.3
pydantic-core==2.14.6
python-dotenv==0.21.1
requests==2.31.0
sniffio==1.3.1
tqdm==4.66.2
typing-extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
yarl==1.9.4
zipp==3.15.0
